requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cssbeautifier==1.15.1
Django==4.2.8
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-phonenumber-field==7.1.0
django-rest-authtoken==2.1.4
django-sass-processor==1.4
django-stubs==5.0.0
django-stubs-ext==5.0.0
django-vite==3.0.1
djangorestframework==3.15.1
djlint==1.34.1
drf-spectacular==0.27.2
EditorConfig==0.12.4
Faker==12.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
inertia-django==0.5.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mixer==7.2.2
pathspec==0.12.1
phonenumbers==8.13.37
pillow==10.3.0
pip-autoremove==0.10.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.1
ruff==0.4.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.4
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
